+++ Dailymlb/app/src/main/java/com/jmp/dailymlb/model/GameStatus.java
@@ -1,0 +1,17 @@
+package com.jmp.dailymlb.model;
+
+public enum GameStatus {
+    Scheduled("예정"),
+    InProgress("진행"),
+    Final("종료"),
+    Suspended("중지"),
+    Postponed("연기"),
+    Canceled("취소");
+    private final String status;
+    GameStatus(String status) {
+        this.status = status;
+    }
+    public String getStatus() {
+        return status;
+    }
+}
