+++ Dailymlb/app/src/main/java/com/jmp/dailymlb/presenter/GamesContract.java
@@ -15,6 +15,6 @@ public interface GamesContract {
         void attachView(View view);
         void detachView();
         void getGamesByDate(int year, int month, int day);
-        void getStadium();
+        void getStadiums();
     }
 }
